backend/requirements.txt
gunicorn==20.1.0
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt
django-cors-headers
psycopg2==2.9.8
Pillow